PFP as tenant
