+++ src/tenants/plantfortheplanet/About/components/ClimateNeutrality.tsx
@@ -1,0 +1,20 @@
+import Image from 'react-bootstrap/Image';
+import styles from './../About.module.scss';
+export default function ClimateNeutrality() {
+  return (
+    <section className={styles.YucantanSection}>
+      <Image fluid src={'/static/images/home/ClimageNeutrality.jpg'} />
+      <div className={styles.YucantanSectionText}>
+        <p className={styles.YucantanSectionTextHeader}>
+          Double Climate Neutrality
+        </p>
+        <p className={styles.YucantanSectionTextPara}>
+          We work with you on your climate journey with Carbon Credits
+        </p>
+        <div className={styles.YucantanSectionTextLinkContainer}>
+          <p className={styles.YucantanSectionTextLink}>Learn More</p>
+        </div>
+      </div>
+    </section>
+  );
+}
