+++ src/tenants/salesforce/About/About.module.scss
@@ -1,4 +1,4 @@
-@import './../../../src/theme/theme';
+@import '../../../theme/theme';
 
 .landingSection {
   width: 100vw;
