+++ src/tenants/salesforce/About/About.tsx
@@ -1,0 +1,19 @@
+import TreeCounterSection from '../TreeCounter/TreeCounter';
+import Articles from './components/Articles';
+import Blogs from './components/Blogs';
+import Landing from './components/Landing';
+import LeaderBoard from './components/LeaderBoard';
+import LearnMore from './components/LearnMore';
+
+export default function About() {
+  return (
+    <main>
+      <Landing />
+      <TreeCounterSection />
+      <LearnMore />
+      <LeaderBoard />
+      <Articles />
+      <Blogs />
+    </main>
+  );
+}
