+++ src/tenants/salesforce/TreeCounter/TreeCounter.module.scss
@@ -1,4 +1,4 @@
-@import './../../../src/theme/theme';
+@import '../../../theme/theme';
 
 .treeCounter {
   position: relative;
