--- tenants/tenantConfig.js
@@ -1,7 +1,0 @@
-export const TenantConfig ={
-    Override:true,
-    FontFamily:'OpenSans',
-    AccentColor:'#ddd',
-    TenantID:'ten_Salesforce',
-    Logo:'linktologo'
-}
